+++ tse_tick/__init__.py
@@ -12,7 +12,7 @@ Developed at Keio University, Nakatsuma Seminar.
 # tse_tick/__init__.py
 import polars as pl
 
-__version__ = "0.9.0"
+__version__ = "0.10.0"
 __author__ = "Kazumi Li"
 __email__ = "kaiwenli@keio.jp"
 __license__ = "MIT"